+++ packages/nadle/src/builtin-tasks/index.ts
@@ -6,3 +6,4 @@ export * from "./pnpx-task.js";
 export * from "./exec-task.js";
 export * from "./copy-task.js";
 export * from "./delete-task.js";
+export { type FileSelector, type FileSelection } from "./file-selection.js";